Moved MockRequest to tests/utils.py
